+++ src/pages/country.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'react-i18next'
 import api from '@/config/api'
 import { List, RegionCard } from '@/components/molecules'
 import { SearchBar } from '@/components/molecules'
-import { Panel } from '@/components/atoms'
+import { Panel } from '@/components/molecules'
 import { useViewState } from '@/config/store'
 import { useLocale } from '@/hooks/use-locale'
 import { useMapbox } from '@/hooks/use-mapbox'
